moved table pagination out of the display table and into d3container file to better acess scrolling
